requirements.txt
Flask==3.0.2
Flask_SocketIO==5.3.6
python-socketio==5.11.1
eventlet==0.35.1
gunicorn==21.2.0